requirements/base.txt
attrs==23.1.0
colorama==0.4.6
crispy-bootstrap5==0.6
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.1.0
django-extensions-shell==1.7.4.1
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
filelock==3.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
ua-parser==0.16.1
uritemplate==4.1.1
user-agents==2.2.0
virtualenv==20.23.0
django-json-field==0.5.7

django-tenants==3.5.0
gunicorn==20.1.0
autopep8==2.0.2
black==23.3.0
flake8==6.0.0
pylint==2.17.4
unimport==0.16.0
isort==5.12.0
pytest==7.3.1
pytest-django==4.5.2
django-tenants==3.5.0
openpyxl==3.1.2

django-allow-cidr==0.7.1
django-storages==1.14
boto3==1.28.45